Established the rust + python + wasm project structure. Added a placeholder rust function for testing.
