+++ Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "gulagcleaner"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "gulagcleaner_lib"
+crate-type = ["cdylib"]
+path = "rust/lib.rs"
+[dependencies]
+pyo3 = { version = "0.20.0", features = ["extension-module"] }
