+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include Cargo.toml
+recursive-include rust *.rs
