+++ rust/gulagcleaner/gulagcleaner.rs
@@ -1,0 +1,5 @@
+//Here I should place my gulagcleaner-rust functions, and then call clean_pdf from python_binder.rs
+
+pub fn sum_as_string(a: usize, b: usize) -> String {
+    (a + b).to_string()
+}
