--- setup.cfg
@@ -1,32 +1,0 @@
-[metadata]
-name = gulagcleaner
-version = 0.8.2
-author = YM162
-author_email = david.fontaneda16@gmail.com
-description = Ad removal tool for PDFs written in python.
-long_description = file: README.md
-long_description_content_type = text/markdown
-url = https://github.com/YM162/gulag-cleaner-cli
-project_urls =
-    Bug Tracker = https://github.com/YM162/gulag-cleaner-cli/issues
-classifiers =
-    Programming Language :: Python :: 3
-    License :: OSI Approved :: MIT License
-    Operating System :: OS Independent
-
-[options]
-install_requires =
-    pdfrw>=0.4
-    pikepdf>=5.1.2
-    pdfminer.six>=20220524
-
-packages = find:
-python_requires = >=3.6
-
-[options.packages.find]
-exclude =
-    tests
-
-[options.entry_points]
-console_scripts =
-    gulagcleaner = gulagcleaner.command_line:main
